image_generation/requirements.txt
flask
python-dotenv
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gunicorn
